♻️ refactor: move the Filter dropdown hooks into hooks folders
Each `*.hook.ts` beside a dropdown becomes `hooks/<useX>/<useX>.ts` behind a
`hooks/index.ts` barrel, matching the layout the rest of the filter uses.
