+++ lib/components/Filter/components/BadgeDropdown/hooks/index.ts
@@ -1,0 +1,1 @@
+export * from './useBadgeMultiSelect/useBadgeMultiSelect';
