+++ lib/components/Filter/components/DateFilterDropdown/hooks/useDateFilterDropdown/useDateFilterDropdown.ts
@@ -2,11 +2,11 @@ import { useCallback, useId, useMemo, useState } from 'react';
 
 import { useFilterContext } from '@/components/Filter/contexts';
 
-import { sendOpenFilterEvent } from '../../events';
-import { useFilterDropdownSync } from '../../hooks';
-import { getLocale } from '../../utils';
+import { sendOpenFilterEvent } from '../../../../events';
+import { useFilterDropdownSync } from '../../../../hooks';
+import { getLocale } from '../../../../utils';
 
-import { Props } from './DateFilterDropdown.types';
+import { Props } from '../../DateFilterDropdown.types';
 
 export const useDateFilterDropdown = ({
   onApply,
