+++ lib/components/Filter/components/TimeFilterDropdown/TimeFilterDropdown.tsx
@@ -12,7 +12,7 @@ import { cn } from '@/utils';
 
 import { FilterDropdown } from '../FilterDropdown/FilterDropdown';
 
-import { useTimeFilterDropdown } from './TimeFilterDropdown.hook';
+import { useTimeFilterDropdown } from './hooks';
 import { Props } from './TimeFilterDropdown.types';
 
 export const TimeFilterDropdown: FC<Props> = ({
